+++ src/main/java/de/redsix/pdfcompare/cli/CliArguments.java
@@ -1,21 +1,148 @@
 package de.redsix.pdfcompare.cli;
 
+import de.redsix.pdfcompare.CompareResult;
+import de.redsix.pdfcompare.CompareResultImpl;
+import de.redsix.pdfcompare.PdfComparator;
+import org.apache.commons.cli.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
 import java.util.Optional;
 
-public interface CliArguments {
+public class CliArguments {
 
-    Boolean areAvailable();
+    private static final Logger LOG = LoggerFactory.getLogger(CliArguments.class);
 
-    Boolean isHelp();
+    private static final int EXPECTED_FILENAME_INDEX = 0;
+    private static final int ACTUAL_FILENAME_INDEX = 1;
 
-    Optional<String> getExpectedFile();
+    private static final int EQUAL_DOCUMENTS_RESULT_VALUE = 0;
+    private static final int UNEQUAL_DOCUMENTS_RESULT_VALUE = 1;
+    private static final int ERROR_RESULT_VALUE = 2;
 
-    Optional<String> getActualFile();
+    private static final String OUTPUT_OPTION = "o";
+    private static final String EXCLUSION_OPTION = "x";
+    private static final String HELP_OPTION = "h";
 
-    Optional<String> getExclusionsFile();
+    private final Options options;
+    private CommandLine commandLine;
 
-    Optional<String> getOutputFile();
+    public CliArguments(String[] args) {
+        options = new Options();
+        options.addOption(buildHelpOption());
+        options.addOption(buildOutputOption());
+        options.addOption(buildExclusionOption());
+        process(args);
+    }
 
-    void printHelp();
+    public boolean areAvailable() {
+        return hasFileArguments() || isHelp();
+    }
 
+    private boolean hasFileArguments() {
+        return commandLine.getArgList().size() == 2 && getExpectedFile().isPresent() && getActualFile().isPresent();
+    }
+
+    private boolean isHelp() {
+        return commandLine.hasOption(HELP_OPTION);
+    }
+
+    /*package*/ Optional<String> getExpectedFile() {
+        return getRemainingArgument(EXPECTED_FILENAME_INDEX);
+    }
+
+    /*package*/ Optional<String> getActualFile() {
+        return getRemainingArgument(ACTUAL_FILENAME_INDEX);
+    }
+
+    /*package*/  Optional<String> getExclusionsFile() {
+        return Optional.ofNullable(commandLine.getOptionValue(EXCLUSION_OPTION));
+    }
+
+    /*package*/ Optional<String> getOutputFile() {
+        return Optional.ofNullable(commandLine.getOptionValue(OUTPUT_OPTION));
+    }
+
+    /*package*/ int printHelp() {
+        new HelpFormatter().printHelp("java -jar pdfcompare-x.x.x.jar [EXPECTED] [ACTUAL]", options);
+        return 0;
+    }
+
+    public int execute() {
+        if (isHelp()) {
+            return printHelp();
+        }
+        if (hasFileArguments()) {
+            return compare();
+        }
+        System.out.println("No files or too many files where passed as arguments\n");
+        printHelp();
+        return ERROR_RESULT_VALUE;
+    }
+
+    private int compare() {
+        try {
+            PdfComparator<CompareResultImpl> pdfComparator = new PdfComparator<>(getExpectedFile().get(), getActualFile().get());
+            getExclusionsFile().ifPresent(pdfComparator::withIgnore);
+            CompareResult compareResult = pdfComparator.compare();
+            getOutputFile().ifPresent(compareResult::writeTo);
+            return (compareResult.isEqual()) ? EQUAL_DOCUMENTS_RESULT_VALUE : UNEQUAL_DOCUMENTS_RESULT_VALUE;
+        } catch (IOException ex) {
+            LOG.error(ex.getMessage());
+            return ERROR_RESULT_VALUE;
+        }
+    }
+
+    private Option buildOutputOption() {
+        return Option.builder(OUTPUT_OPTION)
+                .argName("output")
+                .desc("Provide an optional output file for the result")
+                .hasArg(true)
+                .longOpt("output")
+                .numberOfArgs(1)
+                .required(false)
+                .type(String.class)
+                .valueSeparator('=')
+                .build();
+    }
+
+    private Option buildExclusionOption() {
+        return Option.builder(EXCLUSION_OPTION)
+                .argName("exclusions")
+                .desc("Provide an optional file with exclusions")
+                .hasArg(true)
+                .longOpt("exclusions")
+                .numberOfArgs(1)
+                .required(false)
+                .type(String.class)
+                .valueSeparator('=')
+                .build();
+    }
+
+    private Option buildHelpOption() {
+        return Option.builder(HELP_OPTION)
+                .argName("help")
+                .desc("Displays this text and exit")
+                .hasArg(false)
+                .longOpt("help")
+                .numberOfArgs(0)
+                .required(false)
+                .build();
+    }
+
+    private void process(String[] args) {
+        try {
+            commandLine = new DefaultParser().parse(options, args);
+        } catch (ParseException exception) {
+            throw new CliArgumentsParseException(exception);
+        }
+    }
+
+    private Optional<String> getRemainingArgument(int index) {
+        if (commandLine.getArgList().isEmpty() || commandLine.getArgList().size() < index + 1) {
+            return Optional.empty();
+        }
+        return Optional.of(commandLine.getArgList().get(index));
+    }
 }