+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 #### Added
 - CHANGELOG.md added to the project
+- Add "-x/--exclusions"-option to CLI so it's possible to pass an exclusions file on the CLI